backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
gunicorn==23.0.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
sentry-sdk[fastapi]==2.14.0
